Generalize basic decoders between EVM chains
