+++ rotkehlchen/tests/fixtures/blockchain.py
@@ -13,8 +13,10 @@ from rotkehlchen.chain.ethereum.node_inquirer import EthereumInquirer
 from rotkehlchen.chain.ethereum.transactions import EthereumTransactions
 from rotkehlchen.chain.evm.contracts import EvmContracts
 from rotkehlchen.chain.evm.types import NodeName
+from rotkehlchen.chain.optimism.decoding.decoder import OptimismTransactionDecoder
 from rotkehlchen.chain.optimism.manager import OptimismManager
 from rotkehlchen.chain.optimism.node_inquirer import OptimismInquirer
+from rotkehlchen.chain.optimism.transactions import OptimismTransactions
 from rotkehlchen.chain.substrate.manager import SubstrateChainProperties, SubstrateManager
 from rotkehlchen.chain.substrate.types import SubstrateAddress
 from rotkehlchen.constants.assets import A_DOT, A_KSM
@@ -252,6 +254,30 @@ def fixture_optimism_inquirer(
 @pytest.fixture(name='optimism_manager')
 def fixture_optimism_manager(optimism_inquirer):
     return OptimismManager(node_inquirer=optimism_inquirer)
+
+
+@pytest.fixture(name='optimism_transactions')
+def fixture_optimism_transactions(
+        database,
+        optimism_inquirer,
+):
+    return OptimismTransactions(
+        optimism_inquirer=optimism_inquirer,
+        database=database,
+    )
+
+
+@pytest.fixture(name='optimism_transaction_decoder')
+def fixture_optimism_transaction_decoder(
+        database,
+        optimism_inquirer,
+        optimism_transactions,
+):
+    return OptimismTransactionDecoder(
+        database=database,
+        optimism_inquirer=optimism_inquirer,
+        transactions=optimism_transactions,
+    )
 
 
 @pytest.fixture(name='ksm_rpc_endpoint')
